LibWeb: Move scrollable overflow measurement out of LayoutState::commit
Scrollable overflow measurement, the scroll offset re-clamp, and sticky
inset computation lived at the tail of LayoutState::commit(), so they
could only run as part of a full layout pass. Move them into the new
Layout/ScrollableOverflow.{h,cpp} and a Document step that runs right
after commit, so an upcoming invalidation step can re-run the same
measurement without performing layout. No behavior change intended.
